requirements.txt
RPi.GPIO >= 0.7.0
Flask >= 1.1.2
